Fixed bug in Day 18 solution (readme)
To calculate all possibles sums I used Itertools' `tuple_combinations()` to
iterate over all possible pairs of Snailfish numbers. The problem was that
Snailfish addition is *non-commutative*, meaning my answer was only correct if
the input numbers happened to be in a lucky order. To fix this I replaced the
map with a flat_map that calculates the sums of both b+a and a+b.